+++ Package.swift
@@ -23,14 +23,17 @@ let package = Package(
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
         .package(url: "https://github.com/swiftlang/swift-docc-plugin", exact: "1.4.5"),
+        .package(url: "https://github.com/Orange-OpenSource/ouds-ios", exact: "1.1.0-rc.1"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "DeclarationAccessibility",
-            dependencies: [],
-            resources: [.copy("Resources/Assets.xcassets")]),
+            dependencies: [
+                .product(name: "OUDSSwiftUI", package: "ouds-ios"),
+            ],
+            resources: [.copy("Ressources/Assets.xcassets")]),
         .testTarget(
             name: "DeclarationAccessibilityTests",
             dependencies: ["DeclarationAccessibility"]),